stub engine

+++ apps/backend/src/models/domain.rs
@@ -42,7 +42,7 @@ pub struct User {
 #[derive(Debug, Clone)]
 pub struct Token {
     pub ticker: String,
-    pub decimals: i32,
+    pub decimals: u8,
     pub name: String,
 }
 
@@ -51,9 +51,9 @@ pub struct Market {
     pub id: String, // Generated as "base_ticker/quote_ticker"
     pub base_ticker: String,
     pub quote_ticker: String,
-    pub tick_size: i64,     // Minimum price increment in quote atoms
-    pub lot_size: i64,      // Minimum size increment in base atoms
-    pub min_size: i64,      // Minimum order size in base atoms
+    pub tick_size: u128,    // Minimum price increment in quote atoms
+    pub lot_size: u128,     // Minimum size increment in base atoms
+    pub min_size: u64,      // Minimum order size in base atoms
     pub maker_fee_bps: i32, // Maker fee in basis points (0-10000)
     pub taker_fee_bps: i32, // Taker fee in basis points (0-10000)
 }
@@ -63,12 +63,12 @@ pub struct Order {
     pub id: Uuid,
     pub user_address: String,
     pub market_id: String, // Generated as "base_ticker/quote_ticker"
-    pub price: i64,
-    pub size: i64,
+    pub price: u128,
+    pub size: u128,
     pub side: Side,
     pub order_type: OrderType,
     pub status: OrderStatus,
-    pub filled_size: i64,
+    pub filled_size: u128,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
@@ -81,8 +81,8 @@ pub struct Trade {
     pub seller_address: String,
     pub buyer_order_id: Uuid,
     pub seller_order_id: Uuid,
-    pub price: i64,
-    pub size: i64,
+    pub price: u128,
+    pub size: u128,
     pub timestamp: DateTime<Utc>,
 }
 
@@ -90,8 +90,8 @@ pub struct Trade {
 pub struct Balance {
     pub user_address: String,
     pub token_ticker: String,
-    pub amount: i64,
-    pub open_interest: i64,
+    pub amount: u128,
+    pub open_interest: u128,
     pub updated_at: DateTime<Utc>,
 }
 
@@ -99,9 +99,9 @@ pub struct Balance {
 pub struct Candle {
     pub market_id: String,
     pub timestamp: DateTime<Utc>,
-    pub open: i64,
-    pub high: i64,
-    pub low: i64,
-    pub close: i64,
-    pub volume: i64,
+    pub open: u128,
+    pub high: u128,
+    pub low: u128,
+    pub close: u128,
+    pub volume: u128,
 }
